Remove squirrel-lang
Doesn't have a "master" branch, which borks my update process:

	git submodule foreach "git checkout master && git pull --ff-only"

So far I've been taking the extra step to do the same for Squirrel's HEAD
branch, but while it's nothing personal, I'm not even using it so it's going
out.  If you need it just clone it directly from its own repo, and you'll
get the latest version that way, too.